fix: remove playground worker circular dependency on `index.ts`
