+++ packages/playground-preview-worker/src/realish.ts
@@ -1,5 +1,5 @@
 import { z } from "zod";
-import { PreviewError } from ".";
+import { PreviewError } from "./errors";
 
 const APIResponse = <T extends z.ZodTypeAny>(resultSchema: T) =>
 	z.union([
