+++ packages/playground-preview-worker/src/sentry.ts
@@ -1,6 +1,6 @@
 import { Toucan } from "toucan-js";
 import { ZodError } from "zod";
-import { HttpError, ZodSchemaError } from ".";
+import { HttpError, ZodSchemaError } from "./errors";
 
 export function handleException(e: unknown, sentry: Toucan): Response {
 	console.error(e);
